+++ tests/functional/__init__.py
@@ -3,6 +3,7 @@ Each command should be tested with following scenarios of execution:
   - with required parameters only
   - with all parameters used (if possible as some may be mutually exclusive)
   - with --apiKey parameter used
+  - using --optionsFile if available
   - using wrong apiKey - with patched response from API
   - API returned error with message in JSON
   - API returned error without JSON