+++ tests/functional/test_models.py
@@ -303,12 +303,10 @@ class TestModelUpload(object):
     UPDATE_TAGS_RESPONSE_JSON_200 = example_responses.UPDATE_TAGS_RESPONSE
 
     @mock.patch("gradient.api_sdk.clients.http_client.requests.get")
-    @mock.patch("gradient.api_sdk.clients.http_client.requests.put")
     @mock.patch("gradient.api_sdk.clients.http_client.requests.post")
     def test_should_send_post_request_when_models_update_command_was_used_with_basic_options(
-            self, post_patched, put_patched, get_patched):
+            self, post_patched, get_patched):
         post_patched.return_value = MockResponse(self.CREATE_MODEL_V2_REPONSE)
-        put_patched.return_value = MockResponse()
         get_patched.return_value = MockResponse(self.GET_PRESIGNED_URL_RESPONSE)
 
         runner = CliRunner()
@@ -319,33 +317,34 @@ class TestModelUpload(object):
             result = runner.invoke(cli.cli, self.BASE_COMMAND)
 
             assert result.output == self.EXPECTED_STDOUT, result.exc_info
-            post_patched.assert_called_with(self.URL,
-                                            headers=EXPECTED_HEADERS,
-                                            json=None,
-                                            files=None,
-                                            data=None,
-                                            params=self.BASE_PARAMS)
+            post_patched.assert_has_calls([
+                mock.call(self.URL,
+                          headers=EXPECTED_HEADERS,
+                          json=None,
+                          files=None,
+                          data=None,
+                          params=self.BASE_PARAMS),
+                mock.call(self.GET_PRESIGNED_URL_RESPONSE,
+                          headers={"Content-Type": mock.ANY},
+                          json=None,
+                          files=None,
+                          params=None,
+                          data=mock.ANY)
+            ])
             get_patched.assert_called_once_with(self.GET_PRESIGNED_URL,
                                                 headers=EXPECTED_HEADERS,
                                                 params=self.GET_PRESIGNED_URL_PARAMS,
                                                 json=None,
                                                 )
-            put_patched.assert_called_once_with(self.GET_PRESIGNED_URL_RESPONSE,
-                                                headers={"Content-Type": ""},
-                                                json=None,
-                                                params=None,
-                                                data=mock.ANY)
-            assert put_patched.call_args.kwargs["data"].encoder.fields["file"][0] == self.MODEL_FILE
+            assert post_patched.call_args.kwargs["data"].encoder.fields["file"][0] == self.MODEL_FILE
 
             assert EXPECTED_HEADERS["X-API-Key"] != "some_key"
 
     @mock.patch("gradient.api_sdk.clients.http_client.requests.get")
-    @mock.patch("gradient.api_sdk.clients.http_client.requests.put")
     @mock.patch("gradient.api_sdk.clients.http_client.requests.post")
     def test_should_send_post_request_when_models_update_command_was_used_with_all_options(
-            self, post_patched, put_patched, get_patched):
+            self, post_patched, get_patched):
         post_patched.return_value = MockResponse(self.CREATE_MODEL_V2_REPONSE)
-        put_patched.return_value = MockResponse()
         get_patched.return_value = MockResponse(self.GET_PRESIGNED_URL_RESPONSE)
 
         runner = CliRunner()
@@ -356,23 +355,26 @@ class TestModelUpload(object):
             result = runner.invoke(cli.cli, self.COMMAND_WITH_ALL_OPTIONS)
 
             assert result.output == self.EXPECTED_STDOUT, result.exc_info
-            post_patched.assert_called_with(self.URL,
-                                            headers=EXPECTED_HEADERS,
-                                            json=None,
-                                            files=None,
-                                            data=None,
-                                            params=self.ALL_OPTIONS_PARAMS)
+            post_patched.assert_has_calls([
+                mock.call(self.URL,
+                          headers=EXPECTED_HEADERS,
+                          json=None,
+                          files=None,
+                          data=None,
+                          params=self.ALL_OPTIONS_PARAMS),
+                mock.call(self.GET_PRESIGNED_URL_RESPONSE,
+                          headers={"Content-Type": mock.ANY},
+                          json=None,
+                          files=None,
+                          params=None,
+                          data=mock.ANY)
+            ])
             get_patched.assert_called_once_with(self.GET_PRESIGNED_URL,
                                                 headers=EXPECTED_HEADERS,
                                                 params=self.GET_PRESIGNED_URL_PARAMS,
                                                 json=None,
                                                 )
-            put_patched.assert_called_once_with(self.GET_PRESIGNED_URL_RESPONSE,
-                                                headers={"Content-Type": ""},
-                                                json=None,
-                                                params=None,
-                                                data=mock.ANY)
-            assert put_patched.call_args.kwargs["data"].encoder.fields["file"][0] == self.MODEL_FILE
+            assert post_patched.call_args.kwargs["data"].encoder.fields["file"][0] == self.MODEL_FILE
 
             assert EXPECTED_HEADERS["X-API-Key"] != "some_key"
 
@@ -393,23 +395,28 @@ class TestModelUpload(object):
             result = runner.invoke(cli.cli, self.COMMAND_WITH_API_KEY_PARAMETER_USED)
 
             assert result.output == self.EXPECTED_STDOUT, result.exc_info
-            post_patched.assert_called_with(self.URL,
-                                            headers=EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
-                                            json=None,
-                                            files=None,
-                                            data=None,
-                                            params=self.ALL_OPTIONS_PARAMS)
+            post_patched.assert_has_calls([
+                mock.call(self.URL,
+                          headers=EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
+                          json=None,
+                          files=None,
+                          data=None,
+                          params=self.ALL_OPTIONS_PARAMS
+                          ),
+                mock.call(self.GET_PRESIGNED_URL_RESPONSE,
+                          headers={"Content-Type": mock.ANY},
+                          files=None,
+                          json=None,
+                          params=None,
+                          data=mock.ANY)
+            ])
+
             get_patched.assert_called_once_with(self.GET_PRESIGNED_URL,
                                                 headers=EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                                 params=self.GET_PRESIGNED_URL_PARAMS,
                                                 json=None,
                                                 )
-            put_patched.assert_called_once_with(self.GET_PRESIGNED_URL_RESPONSE,
-                                                headers={"Content-Type": ""},
-                                                json=None,
-                                                params=None,
-                                                data=mock.ANY)
-            assert put_patched.call_args.kwargs["data"].encoder.fields["file"][0] == self.MODEL_FILE
+            assert post_patched.call_args.kwargs["data"].encoder.fields["file"][0] == self.MODEL_FILE
 
             assert EXPECTED_HEADERS["X-API-Key"] != "some_key"
 
@@ -431,23 +438,30 @@ class TestModelUpload(object):
             result = runner.invoke(cli.cli, command)
 
             assert result.output == self.EXPECTED_STDOUT, result.exc_info
-            post_patched.assert_called_with(self.URL,
-                                            headers=EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
-                                            json=None,
-                                            files=None,
-                                            data=None,
-                                            params=self.ALL_OPTIONS_PARAMS)
+            post_patched.assert_has_calls([
+                mock.call(
+                    self.URL,
+                    headers=EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
+                    json=None,
+                    files=None,
+                    data=None,
+                    params=self.ALL_OPTIONS_PARAMS
+                ),
+                mock.call(
+                    self.GET_PRESIGNED_URL_RESPONSE,
+                    headers={'Content-Type': mock.ANY},
+                    json=None,
+                    files=None,
+                    params=None,
+                    data=mock.ANY
+                )
+            ])
             get_patched.assert_called_once_with(self.GET_PRESIGNED_URL,
                                                 headers=EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                                 params=self.GET_PRESIGNED_URL_PARAMS,
                                                 json=None,
                                                 )
-            put_patched.assert_called_once_with(self.GET_PRESIGNED_URL_RESPONSE,
-                                                headers={"Content-Type": ""},
-                                                json=None,
-                                                params=None,
-                                                data=mock.ANY)
-            assert put_patched.call_args.kwargs["data"].encoder.fields["file"][0] == self.MODEL_FILE
+            assert post_patched.call_args.kwargs["data"].encoder.fields["file"][0] == self.MODEL_FILE
 
             assert EXPECTED_HEADERS["X-API-Key"] != "some_key"
 
@@ -488,16 +502,24 @@ class TestModelUpload(object):
                 h.write("I'm a model!")
 
             result = runner.invoke(cli.cli, self.BASE_COMMAND_WITH_TAGS)
-
             assert result.output == self.EXPECTED_STDOUT, result.exc_info
-            post_patched.assert_called_with(
-                self.URL,
-                headers=EXPECTED_HEADERS,
-                json=None,
-                files=None,
-                data=None,
-                params=self.BASE_PARAMS,
-            )
+            post_patched.assert_has_calls([
+                mock.call(self.URL,
+                          headers=EXPECTED_HEADERS,
+                          json=None,
+                          files=None,
+                          data=None,
+                          params=self.BASE_PARAMS
+                          ),
+                mock.call(
+                    self.GET_PRESIGNED_URL_RESPONSE,
+                    headers={"Content-Type": mock.ANY},
+                    files=None,
+                    json=None,
+                    params=None,
+                    data=mock.ANY,
+                ),
+            ])
             get_patched.assert_has_calls(
                 [
                     mock.call(
@@ -510,13 +532,6 @@ class TestModelUpload(object):
             )
             put_patched.assert_has_calls(
                 [
-                    mock.call(
-                        self.GET_PRESIGNED_URL_RESPONSE,
-                        headers={"Content-Type": ""},
-                        json=None,
-                        params=None,
-                        data=mock.ANY,
-                    ),
                     mock.call(
                         self.TAGS_URL,
                         headers=EXPECTED_HEADERS,